Make the build a little more configurable.
We introduce a PREFIX variable to indicate the final location.
DESTDIR is still honoured as the install destination.

We also parameterise the locations of kinit, klist, kdestroy, perl
and /etc/krb5/krb5_keytab.conf.

We install krb5_keytabc into $(PREFIX)/libexec to get it out of people's
paths as it is not exactly designed to be called directly.