Improve usage with -A.
We improve the usage when the -A flag is supplied.  Before the
change, -A would ignore command line arguments and proceed to
construct a set of host principals based on DNS.  We separate the
logic out such that this list is constructed only if there are no
command line arguments in line with our regular usage.  We also
only use the DNS-based mechanism to construct instances that are
attached to host principals under the assumption that we cannot
use our existing host principals to decide what host principals to
create/fetch.  And we default to the service host when being called
by root rather than root.  If command line arguments are presented,
we use them instead.