First pass at bootstrapping support in krb5_keytab.
This will be more fully documented elsewhere and later.  For the
purposes of this commit message, let's just say that we now allow
a special target of: bootstrap/RANDOM[@REALM].  If you request this
and you are root, if will create a ``bootstrap id'' which has the
form boostrap/<stuff>@REALM where <stuff> is alphanumeric.  This
can then be ``bound'' to a hostname via krb5_admin.  When requesting
host keys, krb5_keytab will first try to use existing host keys
but in the absence of them, it will search through the bootstrap
id keys to see if one of them will give us our host keys.

This is a first cut, there are a few things left to fix up.