krb5_keytab.conf had faulty File::Find logic.
We re-organise this a bit as File::Find appears to enforce that
the list that it is passed is a list of directories which is a
little anti-intuitive.  And it emits warnings if the directories
listed do not exist which we would like not to see.  So, we make
all of it conditional on existence.